server/requirements.txt
asyncpg==0.28.0
pydantic==2.0.2
python-dotenv==1.0.0
